[Draw ROI] Moved Drawing Class from MainPageController.py to DrawROIWindow.py. Cleaned unused imports.
